pyproject.toml
[tool.poetry]
name = "repoform"
version = "0.1.0"
description = ""
authors = ["Samy Nitsche <samy@nitsche.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
typer = "^0.9.0"
ruamel-yaml = "^0.17.40"
python-gitlab = "^4.1.1"
deepdiff = "^6.7.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.11.0"

[tool.poetry.scripts]
repoform = "repoform.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
